+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fix: UserTokenServer.delete wasn't using the correct key in to the :ets table
 - New: UserTokenServer.delete_for_shop/1 to delete all the tokens for a shop
+- New: Shop helper functions to translate from slug <> myshopify domain, and new to_uri helper function.
+- New: Shopify admin URI helper
 
 ## 0.15.4